Create method to inject vue state and encrypted wallet in the storage before some tests
