+++ .circleci/config.yml
@@ -34,6 +34,8 @@ jobs:
             - v1-dependencies-{{ checksum "package.json" }}
             # fallback to using the latest cache if no exact match is found
             - v1-dependencies-
+      
+      - run: npm install
 
       - run: npm run build
