+++ src/popup/router/pages/Account.vue
@@ -67,7 +67,6 @@ export default {
       this.loading = false;
     });
     this.pollData();
-    console.log(this.account);
   },
   methods: {
     showAllTranactions() {
